pyproject.toml
[project]
name = "pr-sentinel"
version = "0.1.0"
description = "LangGraph-based GitHub PR Review Agent"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "google-genai>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "python-dotenv>=1.1",
    "langgraph-checkpoint-sqlite>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-httpx>=0.35",
    "ruff>=0.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
